#hw-89 track number validation moved to model

+++ api/routers/artists.ts
@@ -2,9 +2,9 @@ import express from "express";
 import Artist from "../models/Artist";
 import {imagesUpload} from "../multer";
 import mongoose from "mongoose";
-import {IArtist} from "../types";
 import auth, {IRequestWithUser} from "../middlewares/auth";
 import permit from "../middlewares/permit";
+import {IArtist} from "../types";
 
 const artistsRouter = express.Router();
 
@@ -40,12 +40,17 @@ artistsRouter.post('/', auth, imagesUpload.single('image') , async (req, res, ne
     }
 });
 
-artistsRouter.delete('/:id', auth, permit('admin'), async (req, res) => {
+artistsRouter.delete('/:id', auth, permit('admin', 'user'), async (req, res) => {
   try {
+    const user = (req as IRequestWithUser).user;
     const artist = await Artist.findById(req.params.id);
 
     if (!artist) {
       return res.status(404).send({error: 'artist not found'});
+    }
+
+    if (artist.user.toString() !== user._id.toString() && artist.isPublished) {
+      return res.status(403).send({ message: 'Permission denied' });
     }
 
     await Artist.findByIdAndDelete(req.params.id);
